% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{extractCourseInfoFromKmlFiles}
\alias{extractCourseInfoFromKmlFiles}
\title{Read golf course (several holes) info from a KML file}
\usage{
extractCourseInfoFromKmlFiles(kmlFiles, elevationCacheFile = NULL)
}
\arguments{
\item{kmlFiles}{the KML file documenting the holes on the course (see documentation for extractHoleInfoFromKmlLayer on file conventions)}

\item{elevationCacheFile}{the path to a file in which to store a cache of elevations, to avoid over-hitting the Google elevation API}
}
\value{
a tibble with a row for each hole documented in the specified file
}
\description{
Extract information about all the holes on a group of golf courses, documented in the specified vector of KML files.
}

